Moved the list of EFI protocol GUIDs with their names to a seperate file.
* include/linux/efi.h: declaration of the list and relevant functions
* kernel/efi_protocol_guids.c: contains the list and search/comparison
                               functions
* kernel/Makefile: added kernel/efi_protocol_guids.c
